+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:hive_flutter/adapters.dart';
+import 'package:paradise/core/models/room_model.dart';
 import 'package:paradise/presentations/routes.dart';
-import 'package:paradise/presentations/screens/DetailRoom_screen.dart';
 import 'package:paradise/presentations/screens/detail_room.dart';
 import 'package:paradise/presentations/screens/splash_screen.dart';
 import 'core/constants/color_palatte.dart';
@@ -19,7 +19,17 @@ Future main() async {
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
-
+  static final List<String> sub = ['roomDetail2', 'roomDetail3', 'roomDetail4'];
+  static final RoomModel abc = RoomModel(
+      roomID: 'P001',
+      PrimaryImage: 'roomDetail1',
+      RoomKindID: 'TYPE1',
+      price: 150000,
+      State: '3 PEOPLE',
+      SubImages: sub,
+      Description:
+          'Ramayana Prambanan is a show that combines dance and drama without dialogue, based on the Ramayana story, it\'s performed near Prambanan Temple on Java Island, Indonesia. Ramayana Prambanan performs since 1961.',
+      maxCapacity: 3);
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -29,7 +39,9 @@ class MyApp extends StatelessWidget {
         primaryColor: ColorPalette.primaryColor,
         scaffoldBackgroundColor: ColorPalette.backgroundColor,
       ),
-      home: RoomDetailScreen(),
+      home: DetailRoom(
+        room: abc,
+      ),
       routes: routes,
       debugShowCheckedModeBanner: false,
     );
